Add tangential_penalty contact formulation ref #6524
Despite much effort, we are still struggling to get good convergence with
strict kinematic enforcement of frictional contact. This adds a new formulation
that does kinematic enforcement in the normal direction, but treats it as a
penalty constraint in the tangential direction. This isn't perfect, but it's by
far the most robust implementation we have yet.  I've successfully used this on
some very large models.