Do not recurse on semi-local elements
Instead of recursing if elements (or nodes) are semilocal this patch
allows recursion on neighbors of active local elements but nothing
beyond that. I have verified that this at least removes the previous
assertion (seg fault) on the attached 3D input file. This is way
to large to turn into a regression test. More work to manufacture a
good test is needed.

closes #7200